requirements.txt
streamlit
pandas
scikit-learn
nltk
pymupdf
python-docx
groq
python-dotenv
sqlalchemy
